RabbitMQ- rename files
