+++ src/main/java/com/sap/ordermanagergreen/util/rabbitMQ/Producer.java
@@ -1,4 +1,4 @@
-package com.sap.ordermanagergreen.util;
+package com.sap.ordermanagergreen.util.rabbitMQ;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.sap.ordermanagergreen.mapper.OrderMapper;
@@ -8,13 +8,13 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.stereotype.Service;
-import static com.sap.ordermanagergreen.util.RabbitMQConfig.*;
+import static com.sap.ordermanagergreen.util.rabbitMQ.RabbitMQConfig.*;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class DefaultExchangeProducer {
+public class Producer {
     private final RabbitTemplate rabbitTemplate;
     ObjectMapper objectMapper = new ObjectMapper();
     public void sendMessageWaitingForCharge(Order order) throws JsonProcessingException {
